added check for tree declaration
possibly problematic in the current solution: When tree degree of tree declaration is verified we do not send back a checked expression. Example: tree <int>t(2+3) has as its degree 2+3 which is an expression. Currently we verify that 2+3 is of type int but we do not return a checked expression. It will be the job of the intermediate.ml to know that they will receive an Ast.expr as the degree.